refactor(mac): split the last two Mac god files
MacSettingsScene.swift, 803 -> 38. It was one host plus six private
tab structs; each tab is now its own file named after its type, which
is how the rest of Platform/Mac is already organised:

  MacAccountSettingsView.swift     237   sign-in, push diagnostics, sync
  MacDeveloperSettingsView.swift   209   DEBUG-only clock override + VM
  MacSidebarSettingsView.swift     170   pin / unpin / reorder features
  MacGeneralSettingsView.swift      76
  MacAppearanceSettingsView.swift   56
  MacAboutSettingsView.swift        40

What's left is the TabView and nothing else — the file now answers
"which tabs, in what order" and stops. The six `// MARK:` dividers are
gone with it; a file named MacAboutSettingsView.swift does not need a
`// MARK: - About` above its only type. `private struct` widened to
internal so the scene can still name them; MacDebugClockViewModel stays
private, since only the Developer tab in its own file touches it.

MacClassTableView.swift, 820 -> 414, into two extensions:

  MacClassTableView+Grid.swift      246   glass card, columns, cells
  MacClassTableView+Editing.swift   192   add / remove / rename / delete

The core keeps the state, the cache reads, the page chrome and the
lifecycle warm. Swift extensions cannot hold stored properties and
`private` is file-scoped, so 25 declarations widened to internal —
computed exactly, not by blanket search-and-replace: I diffed the
symbols each partition declares against the symbols the other two
reference, and widened only the crossings. Everything genuinely local
to the grid (cellView, blockHeight, emptyCell, the label helpers) is
still private.

Widening @State is the part worth flagging, so MacClassTableView.swift
now carries a note about it: internal @State joins the memberwise
initializer, and anything that fed state in that way would defeat
@State's ownership. Nothing does today, and nothing outside these
three files should.

Both splits are pure moves. Verified by comparing the multiset of
non-blank lines before and after: for MacClassTableView the only
disappearances are the 25 `private` declarations, reappearing without
the keyword; for MacSettingsScene, the six `private struct` lines and
the six MARK dividers. Everything else is identical, and the splitter
asserts each source line lands in exactly one output.

tools/check_macos_sources.py passes with the 8 new files added to both
INCLUDED_SOURCE_FILE_NAMES[sdk=macosx*] arrays (270 sources, 155 macOS,
Debug and Release in agreement). macOS and iOS both build clean.
