+++ swift/TigerDuck/Platform/Mac/MacAppearanceSettingsView.swift
@@ -1,0 +1,56 @@
+#if os(macOS)
+import SwiftUI
+
+/// Appearance tab — accent colour swatches and the light/dark preset.
+/// One of the six tabs assembled by `MacSettingsScene`.
+struct MacAppearanceSettingsView: View {
+    @Environment(AppState.self) private var appState
+
+    var body: some View {
+        @Bindable var state = appState
+        Form {
+            Section(String(localized: "settings_accent_color")) {
+                LazyVGrid(columns: Array(repeating: GridItem(.flexible(), spacing: 12), count: 4), spacing: 12) {
+                    ForEach(AppState.themeColors, id: \.hex) { entry in
+                        accentSwatch(hex: entry.hex)
+                    }
+                }
+                .padding(.vertical, 4)
+            }
+
+            Section(String(localized: "desktop_settings_section_schedule")) {
+                Toggle(String(localized: "settings_show_absolute_assignment_time"), isOn: $state.showAbsoluteAssignmentTime)
+            }
+        }
+        .formStyle(.grouped)
+        .padding(20)
+        .frame(maxWidth: .infinity, maxHeight: .infinity)
+    }
+
+    private func accentSwatch(hex: Int) -> some View {
+        let color = Color(hex: UInt(bitPattern: Int(hex)))
+        let isSelected = appState.accentColorHex == hex
+        return Button {
+            appState.accentColorHex = hex
+        } label: {
+            ZStack {
+                Circle()
+                    .fill(color)
+                    .frame(width: 32, height: 32)
+                if isSelected {
+                    Image(systemName: "checkmark")
+                        .font(.system(size: 14, weight: .bold))
+                        .foregroundStyle(.white)
+                }
+            }
+            .padding(4)
+            .background(
+                Circle()
+                    .stroke(isSelected ? color : Color.clear, lineWidth: 2)
+            )
+        }
+        .buttonStyle(.plain)
+    }
+}
+
+#endif
